make max_runtime Integer

+++ app/src/main/java/org/openobservatory/ooniprobe/model/settings/Settings.java
@@ -43,7 +43,7 @@ public class Settings {
 		@SerializedName("geoip_country_path")
 		public String geoip_country_path;
 		@SerializedName("max_runtime")
-		public Float max_runtime;
+		public Integer max_runtime;
 		@SerializedName("no_collector")
 		public boolean no_collector;
 		@SerializedName("save_real_probe_asn")
